india-food-agent-main/india-food-agent-main/india_food_agent/requirements.txt
anthropic>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.18.0
python-dotenv>=1.0.0
fake-useragent>=1.4.0
lxml>=5.1.0
